refactor: move renderPaginatedTableErrorMessage to utils
